% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.R
\name{chisq_pval}
\alias{chisq_pval}
\title{Visualization of p-values for basic hypothesis tests with the \eqn{\chi2} distribution}
\usage{
chisq_pval(chisq_value = 4, df = 1)
}
\arguments{
\item{chisq_value}{The value of a test statistic with the underlying \eqn{\chi2} distribution

(values that are very far away from the mean - roughly more than 4 times the standard deviation - are not recommend to use as the p-value will be approximately 0 or 1 anyways)}

\item{df}{The degree of freedom of the underlying \eqn{\chi2} distribution (only df greater than 0).}
}
\value{
a ggplot2 object displaying the results
}
\description{
Given \eqn{\chi2 ~ \chi2(df)} the function calculates the p-value and visualizes the result as the area under the density function.
Furthermore the mean and the values one and two standard deviations from the mean are highlighted.
}
\details{
The p-value will be calculated by P(X>\eqn{\chi2}) given X ~ \eqn{\chi2}(df).
}
\examples{
chisq_pval()
chisq_pval(chisq_value = 23, df = 15)
}
\author{
Emanuel Sommer
}
