% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dist.R
\name{F_pval}
\alias{F_pval}
\title{Visualization of p-values for basic hypothesis tests with the F distribution}
\usage{
F_pval(F_value = 5, df1 = 3, df2 = 5)
}
\arguments{
\item{F_value}{The value of a test statistic with the underlying F distribution

(values that are very far away from the mean - roughly more than 4 times the standard deviation - are not recommend to use as the p-value will be approximately 0 anyways)}

\item{df1}{The first degree of freedom of the underlying F distribution (only df1 greater than 2).}

\item{df2}{The second degree of freedom of the underlying F distribution (only df2 greater than 4).}
}
\value{
a ggplot2 object displaying the results
}
\description{
Given \eqn{F ~ F(df1,df2)} the function calculates the p-value and visualizes the result as the area under the density function.
Furthermore the mean and the values one and two standard deviations from the mean are highlighted.
}
\details{
The p-value will be calculated by P(X>F) given X ~ F(df1,df2).
}
\examples{
F_pval()
F_pval(F_value = 3, df1 = 5, df2 = 30)
}
\author{
Emanuel Sommer
}
