% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_pval}
\alias{norm_pval}
\title{Visualization of p-values for basic hypothesis tests with the normal distribution}
\usage{
norm_pval(z_value = 0, mean = 0, sd = 1, direction = c("extreme",
  "less", "greater", "both"))
}
\arguments{
\item{z_value}{The value of a test statistic with the underlying normal distribution}

\item{mean}{The mean of the underlying normal distribution (default is 0).}

\item{sd}{The standard deviation of the underlying normal distribution (default is 1).}

\item{direction}{The 'direction' of the test with respect to z:
\describe{
 \item{extreme}{The p-value will be calculated using \eqn{min(P(X \leq z),P(X \geq z)) with X~N(\mu,\sigma^2)} }
 \item{less}{The p-value will be calculated using \eqn{P(X \leq z) with X~N(\mu,\sigma^2)}}
 \item{greater}{The p-value will be calculated using \eqn{P(X \geq z) with X~N(\mu,\sigma^2)}}
 \item{both}{The p-value will be calculated using \eqn{2*min(P(X \leq z),P(X \geq z)) with X~N(\mu,\sigma^2)} }
 }
 So for the first three options a one sided hypothesis gets tested and for the last one a two sided hypothesis is tested.}
}
\value{
a ggplot2 object displaying the results
}
\description{
Given \eqn{z~N(\mu,\sigma^2)} the function calculates the p-value and visualizes the result as the area under the density function.
}
\examples{
norm_pval(-2)
norm_pval(z_value = 100, mean = 120, sd = 20, direction = "both")
}
\author{
Emanuel Sommer
}
