% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_dist.R
\name{t_pval}
\alias{t_pval}
\title{Visualization of p-values for basic hypothesis tests with the student-t distribution}
\usage{
t_pval(T_value = 0, df, direction = c("extreme", "less", "greater",
  "both"))
}
\arguments{
\item{T_value}{The value of a test statistic with the underlying student-t distribution}

\item{df}{The degree of freedom of the underlying student-t distribution (only df greater than 2).}

\item{direction}{The 'direction' of the test with respect to T:
\describe{
 \item{extreme}{The p-value will be calculated using \eqn{min(P(X \le T),P(X \ge T)) with X~t(df)} }
 \item{less}{The p-value will be calculated using \eqn{P(X \le T) with X~t(df)}}
 \item{greater}{The p-value will be calculated using \eqn{P(X \ge T) with X~t(df)}}
 \item{both}{The p-value will be calculated using \eqn{2*min(P(X \le T),P(X \ge T)) with X~t(df)} }
 }
 So for the first three options a one sided hypothesis gets tested and for the last one a two sided hypothesis is tested.}
}
\value{
a ggplot2 object displaying the results
}
\description{
Given \eqn{T~t(df)} the function calculates the p-value and visualizes the result as the area under the density function.
}
\examples{
t_pval(-2,df=20)
t_pval(T_value = 1, df=10, direction = "both")
}
\author{
Emanuel Sommer
}
